Add prettier script and reorganize files
