+++ package.json
@@ -19,8 +19,9 @@
   },
   "scripts": {
     "lint": "eslint --fix --ext .ts,.tsx ./src",
-    "start": "npm run lint && react-scripts start",
-    "build": "npm run lint && react-scripts build",
+    "format": "prettier --write \"{src,public}/**/*.{js,jsx,ts,tsx,html,css,json}\"",
+    "start": "npm run lint && npm run format && react-scripts start",
+    "build": "npm run lint && npm run format && react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
