+++ public/index.html
@@ -2,14 +2,13 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" /> <!-- TODO: replace logo -->
+    <!-- TODO: replace logo -->
+    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
-    <meta
-      name="description"
-      content="Joe Prado Portfolio"
-    />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" /> <!-- TODO: replace logo -->
+    <meta name="description" content="Joe Prado Portfolio" />
+    <!-- TODO: replace logo -->
+    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <!--
       manifest.json provides metadata used when your web app is installed on a
       user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
