+++ src/App.tsx
@@ -1,11 +1,11 @@
 import styled from '@emotion/styled';
 import { Outlet } from 'react-router-dom';
-import AppHeader from './AppHeader';
+import Header from './components/header/Header';
 
 function App() {
   return (
     <Container>
-      <AppHeader />
+      <Header />
       <ScrollContainer>
         <Main>
           <Outlet />
